+++ test/chex/piece/king_test.exs
@@ -120,4 +120,107 @@ defmodule Chex.Piece.KingTest do
 
     assert moves == []
   end
+
+  describe "castling" do
+    test "white king can castle" do
+      {:ok, game} = Chex.Game.new("r3k2r/8/8/8/8/8/8/R3K2R w KQkq - 0 1")
+      moves = Chex.Piece.King.possible_moves(:white, {:e, 1}, game)
+
+      assert {:g, 1} in moves
+      assert {:c, 1} in moves
+    end
+
+    test "black king can castle" do
+      {:ok, game} = Chex.Game.new("r3k2r/8/8/8/8/8/8/R3K2R b KQkq - 0 1")
+      moves = Chex.Piece.King.possible_moves(:black, {:e, 8}, game)
+
+      assert {:g, 8} in moves
+      assert {:c, 8} in moves
+    end
+
+    test "white cannot castle when blocked by own piece" do
+      {:ok, game} = Chex.Game.new("r3k2r/8/8/8/8/8/8/R1N1KB1R w KQkq - 0 1")
+      moves = Chex.Piece.King.possible_moves(:white, {:e, 1}, game)
+
+      refute {:g, 1} in moves
+      refute {:c, 1} in moves
+    end
+
+    test "black cannot castle when blocked by own piece" do
+      {:ok, game} = Chex.Game.new("r1n1kb1r/8/8/8/8/8/8/R1N1KB1R b KQkq - 0 1")
+      moves = Chex.Piece.King.possible_moves(:black, {:e, 8}, game)
+
+      refute {:g, 8} in moves
+      refute {:c, 8} in moves
+    end
+
+    test "white cannot castle when blocked by enemy piece" do
+      {:ok, game} = Chex.Game.new("r3k2r/8/8/8/8/8/8/R1n1Kb1R w KQkq - 0 1")
+      moves = Chex.Piece.King.possible_moves(:white, {:e, 1}, game)
+
+      refute {:g, 1} in moves
+      refute {:c, 1} in moves
+    end
+
+    test "black cannot castle when blocked by enemy piece" do
+      {:ok, game} = Chex.Game.new("r1N1kB1r/8/8/8/8/8/8/R1N1KB1R b KQkq - 0 1")
+      moves = Chex.Piece.King.possible_moves(:black, {:e, 8}, game)
+
+      refute {:g, 8} in moves
+      refute {:c, 8} in moves
+    end
+
+    test "cannot castle when in check" do
+      {:ok, game} = Chex.Game.new("r3k2r/8/8/1Q6/8/8/8/4K3 b kq - 0 1")
+      game = %{game | check: :black}
+      moves = Chex.Piece.King.possible_moves(:black, {:e, 8}, game)
+
+      refute {:g, 8} in moves
+      refute {:c, 8} in moves
+    end
+
+    test "can castle when rook is attacked" do
+      {:ok, game} = Chex.Game.new("4k3/6q1/8/8/8/8/8/R3K2R w KQ - 0 1")
+      moves = Chex.Piece.King.possible_moves(:white, {:e, 1}, game)
+
+      assert {:c, 1} in moves
+    end
+
+    test "cannot castle into check" do
+      {:ok, game} = Chex.Game.new("4k3/6q1/8/8/8/8/8/R3K2R w KQ - 0 1")
+      moves = Chex.Piece.King.possible_moves(:white, {:e, 1}, game)
+
+      refute {:g, 1} in moves
+    end
+
+    test "cannot castle through attacked squares" do
+      {:ok, game} = Chex.Game.new("4k3/8/8/8/6q1/8/8/R3K2R b KQ - 0 1")
+      moves = Chex.Piece.King.possible_moves(:white, {:e, 1}, game)
+
+      refute {:g, 1} in moves
+      refute {:c, 1} in moves
+    end
+
+    test "cannot queenside castle when rook is blocked by own piece" do
+      {:ok, game} = Chex.Game.new("r3k2r/8/8/8/8/8/8/RN2K2R b KQkq - 0 1")
+      moves = Chex.Piece.King.possible_moves(:white, {:e, 1}, game)
+
+      refute {:c, 1} in moves
+    end
+
+    test "cannot queenside castle when rook is blocked by enemy piece" do
+      {:ok, game} = Chex.Game.new("r3k2r/8/8/8/8/8/8/Rn2K2R b KQkq - 0 1")
+      moves = Chex.Piece.King.possible_moves(:white, {:e, 1}, game)
+
+      refute {:c, 1} in moves
+    end
+
+    test "attacking b file does not prevent queenside castling" do
+      # Averbakh vs. Purdy, 1960
+      {:ok, game} = Chex.Game.new("r3kb1r/p5pp/2p1bp2/4p3/2P5/2P5/P2NPP1P/1RB1K2R w Kkq - 0 1")
+      moves = Chex.Piece.King.possible_moves(:black, {:e, 8}, game)
+
+      assert {:c, 8} in moves
+    end
+  end
 end